customercare/serializers.py
from rest_framework import serializers
from customercare.models import Comment, COMMENT_TYPE, Feedback


class CommentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Comment
        fields = '__all__'


class FeedbackSerializer(serializers.ModelSerializer):
    class Meta:
        model = Feedback
        fields = '__all__'
