requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.3
Markdown==3.3.6
psycopg2-binary==2.9.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
zipp==3.7.0
